+++ src/pages/dar_application/ResearcherInfo.js
@@ -2,7 +2,7 @@ import { useState, useEffect} from 'react';
 import { Alert } from '../../components/Alert';
 import { Link } from 'react-router-dom';
 import { a, div, fieldset, h, h2, h3, h4, span } from 'react-hyperscript-helpers';
-import ERACommons from '../../components/ERACommons';
+import ERACommons from '../../components/eRACommons';
 import CollaboratorList from './collaborator/CollaboratorList';
 import { isEmpty, isNil, get } from 'lodash/fp';
 import { FormField, FormValidators, FormFieldTypes } from '../../components/forms/forms';
