+++ src/pages/user_profile/ResearcherStatus.js
@@ -1,5 +1,5 @@
 import React, { useEffect, useState } from 'react';
-import ERACommons from '../../components/ERACommons';
+import ERACommons from '../../components/eRACommons';
 import { Notifications } from '../../libs/utils';
 import { User } from '../../libs/ajax';
 import { isNil } from 'lodash';
